some cleanup and a stab at fixing ffi by moving error handling over
