+++ libs/core/Cargo.toml
@@ -28,8 +28,6 @@ raqote = { version = "0.8.0", default-features = false }
 reqwest = { version = "0.11.1", default-features = false, features = ["blocking", "json"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.44"
-strum = "0.19.5"
-strum_macros = "0.19.4"
 uuid = { version = "1.2.2", features = ["v4", "serde"] }
 itertools = "0.10.1"
 backtrace = "0.3"
