+++ libs/core_external_interface/Cargo.toml
@@ -18,4 +18,5 @@ lockbook-core = { path = "../core" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.44"
 strum = "0.19.5"
+strum_macros = "0.19.4"
 tracing = "0.1.5"
