+++ libs/core_external_interface/src/c_interface.rs
@@ -487,27 +487,6 @@ pub unsafe extern "C" fn delete_pending_share(id: *const c_char) -> *const c_cha
     })
 }
 
-/// # Safety
-///
-/// Be sure to call `release_pointer` on the result of this function to free the data.
-#[no_mangle]
-pub unsafe extern "C" fn validate() -> *const c_char {
-    c_string(match static_state::get() {
-        Ok(core) => translate(
-            // Map any warnings to Strings as well as any errors using Debug impl text.
-            core.validate()
-                .map(|warnings| {
-                    warnings
-                        .into_iter()
-                        .map(|w| w.to_string())
-                        .collect::<Vec<String>>()
-                })
-                .map_err(|err| err.to_string()),
-        ),
-        e => translate(e.map(|_| ())),
-    })
-}
-
 // FOR INTEGRATION TESTS ONLY
 /// # Safety
 ///
